public interface Connection extends Publisher, Subscriber, AutoCloseable
Should be helpful when writing Java based Connectors by enforcing the
implementation of all the available methods to publish & to subscribe.
